+++ packages/core/src/index.ts
@@ -28,4 +28,5 @@ export * from './application/use-cases/media';
 export * from './application/use-cases/publishing';
 export * from './application/use-cases/webhooks';
 export * from './infra/crypto/aes-gcm.service';
+export * from './infra/storage/local.storage';
 export * from './infra/media/sniff';